Create 2D Arrays (hourglass) problem
Here is an answer to one of the more complicated 2d array problems in swift. Its hard to find swift specific solutions for problems on hackrank so I wanted to give more information to the first initial link on google I could find. Also, my solution uses a recursive method